sky/requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-chess>=1.10.0
groq>=0.9.0
